feat: updates

+++ uninstall.sh
@@ -1,32 +1,43 @@
 #!/bin/bash
 
+set -e
+trap 'echo -e "\033[0;31mError: Command on line $LINENO failed.\033[0m"' ERR
+
 # Source the configuration file
 source config.cfg
 
-# Parameters
+# Automatically detect the username
+USERNAME=$(whoami)
 
-# Stop and remove the systemd service
-echo "Stopping the service..."
-sudo systemctl stop microscope-photo-server
-sudo systemctl disable microscope-photo-server
-echo "Removing the service..."
-sudo rm /etc/systemd/system/microscope-photo-server.service
-sudo systemctl daemon-reload
-sudo systemctl reset-failed
+# Check if the service is active and running
+if sudo systemctl is-active --quiet microscope-photo-server; then
+    # Stop and disable the service
+    sudo systemctl stop microscope-photo-server
+    sudo systemctl disable microscope-photo-server
+fi
 
-# Remove udev rule and permissions
-echo "Removing udev rules..."
-sudo rm /etc/udev/rules.d/40-camera.rules
-sudo udevadm control --reload-rules
-sudo udevadm trigger
+# Check if the service file exists
+if [ -e "/etc/systemd/system/microscope-photo-server.service" ]; then
+    # Remove the systemd service file
+    sudo rm /etc/systemd/system/microscope-photo-server.service
 
-# Remove the Samba share configuration
-echo "Removing Samba share..."
+    # Reload systemd units
+    sudo systemctl daemon-reload
+
+    echo "Service uninstalled."
+else
+    echo "Service file not found. No action taken."
+fi
+
+# Remove the Samba configuration
 sudo sed -i "/\[$SHARE_NAME\]/,/^$/d" /etc/samba/smb.conf
 sudo systemctl restart smbd
 
-# Remove the cloned repository
-echo "Removing repository directory..."
-sudo rm -rf $REPO_DIR
+# Remove the udev rule
+sudo rm /etc/udev/rules.d/40-camera.rules
+sudo usermod -G $USERNAME -d /home/$USERNAME $USERNAME  # Revert user to original group and home directory
+
+# Remove installed files
+sudo rm -rf $INSTALL_DIR
 
 echo "Uninstallation complete!"
